put c api loader into distinct include
